% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{circle}
\alias{circle}
\title{A circle}
\format{A data frame with 2500 rows and 3 variables:
\describe{
  \item{x}{x coordinate of the circle}
  \item{y}{y coordinate of the circle}
  \item{group}{value for use in spatial methods}
}}
\source{
Robert Garrett
}
\usage{
circle
}
\description{
A code-generated circle that intersects the points dataset.
}
\keyword{datasets}
