% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_map.R
\name{mask_map}
\alias{mask_map}
\title{Invert a sp polygon}
\usage{
mask_map(map, width = 5, piece = 1)
}
\arguments{
\item{map}{SpatialPolygonsDataFrame with lat and long}

\item{width}{mask border size}

\item{piece}{if your SPDF has pieces choose which one to mask}
}
\description{
invert a map you want to use voronoi_polygon() on to hide voronoi lines
}
\examples{
ggplot+geom_polygon(mask_map(map),aes(x=long,y=lat))
mask_map()
}
\keyword{mask}
