% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oxford_bikes}
\alias{oxford_bikes}
\title{Bike Rack Locations in Oxford, Ohio.}
\format{
A data frame with 205 rows and 3 variables:
\describe{
  \item{x}{longitude coordinate}
  \item{y}{latitude coordinate}
  \item{name}{location name, if recorded}
}
}
\source{
Madeline Maurer (Miami University Department of Geography); 
Justin Hopkins, Dr. Helaine Alessio and Amanda Meiering (Miami University Department of Kinesiology)
}
\usage{
oxford_bikes
}
\description{
A dataset with the location of bike racks in Oxford, Ohio. Includes Miami University's campus and Uptown Oxford. 
GPS coordinates were measured for each bike rack via handheld GPS receivers. 
Off-campus data were gathered for a Miami University GIS class project and 
On-campus data were gathered for a Miami University Kinesiology research project.
}
\keyword{datasets}
