% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oxford_map}
\alias{oxford_map}
\title{Google Map of Oxford, Ohio.}
\format{
Large ggmap object
}
\source{
Google Maps static map api (accessed from 'ggmap' package)
}
\usage{
oxford_map
}
\description{
A ggmap object containing a static google map image of Oxford, Ohio. Zoomed in to show the Uptown Oxford and Miami University campus areas.
}
\keyword{datasets}
