% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{points}
\alias{points}
\title{Random Points.}
\format{A data frame with 200 rows and 3 variables:
\describe{
  \item{x}{random integer between 1 and 100}
  \item{y}{random integer between 1 and 100}
  \item{fill}{distance of point from (50,50)}
}}
\source{
Robert Garrett
}
\usage{
points
}
\description{
A random set of points intended to demonstrate geom_voronoi().
}
\keyword{datasets}
