% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi_path.R
\name{voronoi_path}
\alias{voronoi_path}
\title{Voronoi tesselation through geom_path}
\usage{
voronoi_path(x = sample(1:100, 100), y = sample(1:100, 100),
  color = "black", size = 0.5, linetype = 1, alpha = NA)
}
\arguments{
\item{x}{numeric vector (for example longitude)}

\item{y}{numeric vector (for example latitude)}

\item{color}{line color (see ggplot2)}

\item{size}{line size (see ggplot2)}

\item{linetype}{see ggplot2}

\item{alpha}{see ggplot2}
}
\description{
Wrapper on geom_path using deldir
}
\examples{
ggplot()+voronoi_path(x=1:5,y=sample(1:5,5),color="darkblue",size=1)
voronoi_path()
}
\keyword{voronoi}
