% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi_polygon.R
\name{voronoi_polygon}
\alias{voronoi_polygon}
\title{Voronoi Diagram from Data Frame}
\usage{
voronoi_polygon(data, x = "x", y = "y", outline = NULL,
  data.frame = FALSE)
}
\arguments{
\item{x}{numeric vector (for example longitude)}

\item{y}{numeric vector (for example latitude)}

\item{outline}{data.frame with first column x/longitude, second column y/latitude, and a group column denoting islands or pieces}

\item{data.frame}{output as data.frame? You will lose information if you do this. For use in StatVoronoi.}
}
\description{
Create a Voronoi diagram for analysis or plotting with geom_polygon
}
\examples{
ggplot()+voronoi_polygon(data=points,x="x",y="y",outline=circle)
voronoi_polygon()
}
\keyword{choropleth}
\keyword{voronoi,}
