% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_voronoi.R
\name{fortify_voronoi}
\alias{fortify_voronoi}
\title{Create Data Frame from voronoi_polygons output}
\usage{
fortify_voronoi(vor_spdf)
}
\arguments{
\item{vor_spdf}{Output from voronoi_polygons (Spatial Polygons Data Frame)}
}
\description{
Use this before attempting to plot with geom_polygon
}
\examples{
fortify_voronoi(vor_spdf)
vor_spdf = voronoi_polygon(data=points,x="x",y="y",outline=circle)
vor_df = fortify_voronoi(vor_spdf)
ggplot(vor_df)+geom_polygon(aes(x=x,y=y,fill=fill,group=group))

}
\keyword{choropleth}
\keyword{voronoi,}
