% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oxford_bikes}
\alias{oxford_bikes}
\title{Bike Rack Locations in Oxford, Ohio.}
\format{A data frame with 205 rows and 3 variables:
\describe{
  \item{x}{longitude coordinate}
  \item{y}{latitude coordinate}
  \item{name}{location name, if recorded}
}}
\source{
Madeline Maurer, Justin Hopkins, Dr. Helaine Alessio, Amanda Meiering
}
\usage{
oxford_bikes
}
\description{
A dataset with the location of bike racks in Oxford, Ohio. Includes Miami University's campus and Uptown Oxford.
}
\keyword{datasets}
