% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi_polygon.R
\name{voronoi_polygon}
\alias{voronoi_polygon}
\title{Voronoi chloropleth through geom_polygon}
\usage{
voronoi_polygon(x = sample(1:100, 100), y = sample(1:100, 100),
  color = "black", fill = "white", size = NA, linetype = 1,
  alpha = NA)
}
\arguments{
\item{x}{numeric vector (for example longitude)}

\item{y}{numeric vector (for example latitude)}

\item{color}{line color (see ggplot2)}

\item{fill}{numeric vector of polygon colors (map chloropleth variable here)}

\item{size}{line size (see ggplot2)}

\item{linetype}{see ggplot2}

\item{alpha}{see ggplot2}
}
\description{
Wrapper on geom_polygon using deldir
}
\examples{
ggplot()+voronoi_polygon(x=1:5,y=sample(1:5,5),fill=1:5,color=NA,size=1)
voronoi_polygon()
}
\keyword{chloropleth}
\keyword{voronoi,}
