% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi_polygon.R
\name{voronoi_polygon}
\alias{voronoi_polygon}
\title{Voronoi Diagram from Data Frame}
\usage{
voronoi_polygon(data, x = "x", y = "y", outline = NULL,
  data.frame = FALSE)
}
\arguments{
\item{data}{data.frame containing a set of points to make voronoi regions and any additional desired columns.}

\item{x}{numeric vector (for example longitude).}

\item{y}{numeric vector (for example latitude).}

\item{outline}{data.frame with first column x/longitude, second column y/latitude, and a group column denoting islands or pieces.}

\item{data.frame}{output as data.frame? You will lose information if you do this. For use in StatVoronoi.}
}
\description{
Create a Voronoi diagram for analysis or plotting with geom_polygon.
}
\examples{
vor_spdf = voronoi_polygon(data=points,x="x",y="y",outline=circle)
vor_df = fortify_voronoi(vor_spdf)

library(ggplot2)
ggplot(vor_df)+
    geom_polygon(aes(x=x,y=y,fill=fill,group=group))
}
\keyword{choropleth}
\keyword{voronoi,}
